+++ app/page.tsx
@@ -1,40 +1,15 @@
 "use client";
 import styles from "./page.module.css";
-import type Asagohan from "@/app/types/Asagohan";
 import Image from "next/image";
 import Link from "next/link";
 import Header from "./components/Header";
-import { useEffect, useState } from "react";
-
-const getTodayAsagohans = async (userID: string): Promise<Asagohan[]> => {
-  const res = await fetch(`http://localhost:3000/api/asagohans/${userID}`);
-  if (!res.ok) {
-    throw new Error("Failed to fetch data");
-  }
-  const asagohans = await res.json();
-  return asagohans.data;
-};
+import { useState } from "react";
+import useTodayAsagohans from "@/app/hooks/useTodayAsagohans";
 
 export default function Home() {
   const [userID, setUserID] = useState("b2113406-aaaf-43bc-a32c-a5cc003506d7");
-  const [asagohans, setAsagohans] = useState<Asagohan[]>([]);
-  const [loading, setLoading] = useState(false);
-  console.log(asagohans);
-
-  useEffect(() => {
-    setLoading(true);
-    getTodayAsagohans(userID)
-      .then((asagohans) => {
-        setAsagohans(asagohans);
-      })
-      .catch((error) => {
-        console.error(error);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  }, [userID]);
-
+  const { asagohans, todayAsagohansFetching } = useTodayAsagohans(userID);
+  console.log(asagohans, todayAsagohansFetching);
   return (
     <div className={styles.page}>
       <Header>
